The form 'Scanned Notes' is modified to upload files. The errors in the form are removed. Any type of file can be uploaded,earlier only it was allowing images.
The files uploaded gets saved in the path \sites\default\documents\pid\encounter\
The present document management system is used to store the details of the uploaded file.Table 'documents' stores the details.
A category 'Scanned Encounter Notes' is added.
The table 'form_scanned_notes' stores the note and the document id.New field is added to store the id.

Selecting the encounter it displays the file along with the form listing.
Can change the file in the edit screen.

The Past Encounter List(Clinical View) shows the files against the form 'Scanned Notes' in the corresponding encounters.
This is the same with the Report link from the demo screen.Report can the generated by ticking the check box.

More over the documents uploaded at the patient level through the Documents link from demo are not affected.
They come in the Clinical View and the Report link as usual.

Signed-off-by: Paul Simon <paul@zhservices.com>