Module installer envisions giving the community the option to install enhancements as compartmentalized code in order to help maintain and manage installations.  Thus enabling contributors to release features as modules which users can opt to use without hindering the basic structure and direction and the common project.
This enables developers to release features without effecting the core structure of the OEMR project and mitigating risks only to their individual installations. This would enable different installations to have multiple solutions on the same installation and implement solutions to real business problems faced by their practices only ie users will have the ability to implement a calendar which has a certain feature or look that is not viable to any other practice.

This gives the community a flavour of multiple solutions; of which some can be eventually be made of the main source code when the project leads have time to complete their testing and review. This is could also lead to vendors allowing third party modules made available to individual users for a price thus increasing the acceptance of the product leading to similar scenario like the app marketplace

Signed-off-by: Jacob T.Paul <jacob@zhservices.com>